zadania/Z0202test/requirements.txt
mcp>=1.2.0
python-dotenv>=1.0.0
# Add your LLM SDKs and dependencies here (e.g., openai, anthropic)
requests
